+++ mlxtend/feature_selection/column_selector.py
@@ -95,7 +95,7 @@ class ColumnSelector(BaseEstimator):
                 )
         else:
             t = X[:, self.cols]
-            
+
         if t.shape[-1] == 1 and self.drop_axis:
             t = t.reshape(-1)
         if len(t.shape) == 1 and not self.drop_axis: