+++ docs/sources/CHANGELOG.md
@@ -25,7 +25,7 @@ The CHANGELOG for the current development version is available at
 
 ##### Bug Fixes
 
-- Fix unreable labels in `plot_confusion_matrix` for imbalanced datasets if `show_absolute=True` and `show_normed=True`.
+- Fix unreadable labels in `plot_confusion_matrix` for imbalanced datasets if `show_absolute=True` and `show_normed=True`. ([#504](https://github.com/rasbt/mlxtend/pull/504))
 
 - Raises a more informative error if a `SparseDataFrame` is passed to `apriori` and the dataframe has integer column names that don't start with `0` due to current limitations of the `SparseDataFrame` implementation in pandas. ([#503](https://github.com/rasbt/mlxtend/pull/503))